requirements.txt
absl-py==0.7.0
aenum==2.1.2
argh==0.26.2
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
backcall==0.1.0
backports.functools-lru-cache==1.5
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
decorator==4.4.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
grpcio==1.16.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
jaraco.functools==2.0
jedi==0.13.3
Jinja2==2.10.1
Markdown==3.0.1
MarkupSafe==1.1.1
mock==2.0.0
more-itertools==7.1.0
parso==0.3.4
pathtools==0.1.2
pbr==5.1.3
pickleshare==0.7.5
Pillow==6.0.0
portend==2.5
prompt-toolkit==2.0.9
protobuf==3.6.1
pycosat==0.6.3
Pygments==2.3.1
pyOpenSSL==18.0.0
pyreadline==2.1
PySocks==1.6.8
python-dateutil==1.5
pytz==2018.9
PyYAML==5.1.1
reportlab==3.5.23
requests==2.21.0
ruamel-yaml==0.15.46
six==1.12.0
SQLAlchemy==1.3.5
tempora==1.14.1
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
WTForms==2.2.1
zc.lockfile==1.4
